fix: create components routing module, fix scss imports
