+++ client/src/app/app-routing.module.ts
@@ -1,30 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AuthGuard } from './auth/auth.guard';
-import { CategoriesComponent } from './components/categories/categories.component';
-import { LoginComponent } from './components/login/login.component';
-import { MainComponent } from './components/main/main.component';
-import { NotFoundComponent } from './components/not-found/not-found.component';
-import { StatisticsComponent } from './components/statistics/statistics.component';
 
-const routes: Routes = [
-  { path: 'login', component: LoginComponent },
-  {
-    path: '',
-    component: MainComponent,
-    canActivate: [AuthGuard],
-    children: [
-      {
-        path: 'categories',
-        component: CategoriesComponent,
-      },
-      { path: 'statistics', component: StatisticsComponent },
-    ],
-  },
-
-  { path: '', pathMatch: 'full', redirectTo: '/' },
-  { path: '**', component: NotFoundComponent },
-];
+const routes: Routes = [{ path: '', pathMatch: 'full', redirectTo: '/' }];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
