+++ client/src/app/components/accounts/accounts-item/accounts-item.component.scss
@@ -1,4 +1,4 @@
-@import "../../shared/scss/variables";
+@import "../../../shared/scss/variables";
 
 .account__item {
   display: flex;
