+++ client/src/app/components/components-routing.module.ts
@@ -1,0 +1,32 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { AuthGuard } from '../auth/auth.guard';
+import { CategoriesComponent } from './categories/categories.component';
+import { LoginComponent } from './login/login.component';
+import { MainComponent } from './main/main.component';
+import { NotFoundComponent } from './not-found/not-found.component';
+import { StatisticsComponent } from './statistics/statistics.component';
+
+const routes: Routes = [
+  { path: 'login', component: LoginComponent },
+  {
+    path: '',
+    component: MainComponent,
+    canActivate: [AuthGuard],
+    children: [
+      {
+        path: 'categories',
+        component: CategoriesComponent,
+      },
+      { path: 'statistics', component: StatisticsComponent },
+    ],
+  },
+  { path: '**', component: NotFoundComponent },
+];
+
+@NgModule({
+  declarations: [],
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule],
+})
+export class ComponentsRoutingModule {}
