+++ client/src/app/components/footer/footer.component.scss
@@ -1,4 +1,4 @@
-@import "../shared/scss/variables";
+@import "../../shared//scss/variables";
 
 :host {
   margin-top: auto;
