+++ client/src/app/components/login/login.component.scss
@@ -1,4 +1,5 @@
-@import "../shared/scss/_variables";
+@import "../../shared/scss/variables";
+
 .login {
   display: flex;
   align-items: center;
